+++ extensions/github-codespaces/src/actions/CodespaceActions.tsx
@@ -1,21 +1,11 @@
 import { Action, ActionPanel, Icon } from "@raycast/api";
 import { preferredEditor } from "../preferences";
-import { Codespace } from "../types";
+import { Client, clientNames, Codespace } from "../types";
 import handleDelete from "../methods/handleDelete";
 import ChangeCompute from "../views/ChangeCompute";
 import Rename from "../views/Rename";
 import { handleStop } from "../methods/handleStop";
-
-const OpenWebEditorAction = ({ codespace }: { codespace: Codespace }) => (
-  <Action.OpenInBrowser title="Open on web" url={codespace.web_url} />
-);
-const OpenVSCodeAction = ({ codespace }: { codespace: Codespace }) => (
-  <Action.OpenInBrowser
-    icon={Icon.Code}
-    title="Open in VS Code"
-    url={`vscode://github.codespaces/connect?name=${codespace.name}&windowId=_blank`}
-  />
-);
+import { launchEditor } from "../utils/launchEditor";
 
 function CodespaceActions({
   codespace,
@@ -24,21 +14,23 @@ function CodespaceActions({
   codespace: Codespace;
   onRevalidate: () => void;
 }) {
-  const PreferredAction = preferredEditor
-    ? preferredEditor === "vscode"
-      ? OpenVSCodeAction
-      : OpenWebEditorAction
-    : OpenVSCodeAction;
-  const SecondaryAction = preferredEditor
-    ? preferredEditor === "vscode"
-      ? OpenWebEditorAction
-      : OpenVSCodeAction
-    : OpenWebEditorAction;
   return (
     <ActionPanel>
       <ActionPanel.Section title="Open">
-        <PreferredAction codespace={codespace} />
-        <SecondaryAction codespace={codespace} />
+        <Action
+          title={`Open in ${clientNames[preferredEditor]}`}
+          onAction={() => launchEditor({ codespace })}
+        />
+        {Object.keys(clientNames).map((c) => {
+          if (c === preferredEditor) return null;
+          return (
+            <Action
+              key={c}
+              title={`Open in ${clientNames[c as Client]}`}
+              onAction={() => launchEditor({ codespace, client: c as Client })}
+            />
+          );
+        })}
       </ActionPanel.Section>
       {!codespace.pending_operation && (
         <>