+++ extensions/github-codespaces/src/views/Rename.tsx
@@ -1,6 +1,7 @@
 import { Action, ActionPanel, Form, useNavigation } from "@raycast/api";
 import { Codespace } from "../types";
 import handleRename from "../methods/handleRename";
+import { useState } from "react";
 
 const Rename = ({
   codespace,
@@ -9,6 +10,7 @@ const Rename = ({
   codespace: Codespace;
   onRevalidate: () => void;
 }) => {
+  const [nameError, setNameError] = useState<string | undefined>();
   const { pop } = useNavigation();
   return (
     <Form
@@ -31,6 +33,16 @@ const Rename = ({
         title="Name"
         defaultValue={codespace.display_name || ""}
         placeholder="Select a display name for this codespace"
+        error={nameError}
+        onBlur={(e) => {
+          if (e.target.value?.length == 0) {
+            setNameError("Name is required");
+          } else {
+            if (nameError && nameError.length > 0) {
+              setNameError(undefined);
+            }
+          }
+        }}
       />
     </Form>
   );