+++ extensions/github-codespaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # GitHub Codespaces Changelog
 
+## [Supporting insiders] - 2022-09-22
+
+- Added support for the Insiders build for VS Code
+- Bug fix - prevent empty codespace name
+- Bug fix - prevent permanent loading state when error occurs
+
 ## [Added menubar and stopping] - 2022-09-19
 
 - Added the ability to stop active codespaces